prompt_attack/PromptAttack_LLM/requirements.txt
accelerate
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
anyio==4.0.0
anytree==2.9.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bert-score
bitsandbytes==0.41.1
boto3==1.28.42
botocore==1.31.42
bpemb==0.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.0.4
cloudpickle==2.2.1
cmake==3.27.2
codernitydb3==0.6.0
comet-ml==3.33.10
comm==0.1.4
configobj==5.0.8
conllu==4.5.3
contourpy==1.1.0
cpm-kernels==1.0.11
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.14.5
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.5.1
docker-pycreds==0.4.0
docopt==0.6.2
dulwich==0.21.6
editdistance==0.6.2
einops==0.6.1
everett==3.1.0
executing==1.2.0
fastapi==0.95.1
ffmpy==0.3.1
filelock==3.12.2
fire==0.5.0
flair
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.4.0
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.35
gradio==3.41.2
gradio_client==0.5.0
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hyperopt==0.2.7
idna==3.4
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
Janome==0.5.0
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.4
langchain==0.0.279
langdetect==1.0.9
langsmith==0.0.32
language-tool-python==2.7.1
lemminflect==0.2.3
lit==16.0.6
loguru==0.7.0
lru-dict==1.2.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
mpld3==0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.13
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
ninja==1.11.1
nltk==3.8.1
num2words==0.5.12
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==1.35.13
OpenHowNet==2.0
orjson==3.9.5
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
peft
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pinyin==0.4.0
platformdirs==3.10.0
pptree==3.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==12.0.1
pycld2==0.41
pydantic==1.10.11
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch_revgrad
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
ray==2.6.3
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.5.2
rouge-chinese==1.0.3
rpds-py==0.10.0
s3transfer==0.6.2
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
screen==1.0.1
segtok==1.5.11
semantic-version==2.10.0
sentence-transformers
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
simplejson==3.19.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
sqlitedict==2.1.0
sse-starlette==1.6.5
stack-data==0.6.2
starlette==0.26.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboardX==2.6.2.2
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformer-smaller-training-vocab
transformers==4.31.0
transformers-stream-generator==0.0.4
triton==2.0.0
trl==0.7.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
vllm==0.1.4
wandb==0.15.10
wcwidth==0.2.6
websocket-client==1.3.3
websockets==11.0.3
Wikipedia-API==0.6.0
word2number==1.1
wrapt==1.15.0
wurlitzer==3.0.3
xformers==0.0.21
xxhash==3.3.0
yarl==1.9.2